feat: add habilidades do usuário

+++ frontend/src/app/service/authService.service.ts
@@ -1,16 +1,44 @@
 import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
-import { Injectable } from '@angular/core';
 import { jwtDecode } from 'jwt-decode';
+import { Injectable, PLATFORM_ID, Inject } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({
   providedIn: 'root',
 })
 export class AuthService {
   private API = 'http://localhost:8080';
+  private readonly TOKEN_KEY = 'access_token'; // Nome consistente
   clientId = 'myclientid'; // colocar em variavel de ambiente
   clientSecret = 'myclientsecret'; // colocar em variavel de ambiente
 
-  constructor(private http: HttpClient) {}
+  constructor(
+    private http: HttpClient,
+    @Inject(PLATFORM_ID) private platformId: Object
+  ) {}
+
+  getToken(): string | null {
+    if (isPlatformBrowser(this.platformId)) {
+      return localStorage.getItem(this.TOKEN_KEY);
+    }
+    return null;
+  }
+
+  setToken(token: string): void {
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.setItem(this.TOKEN_KEY, token);
+    }
+  }
+
+  removeToken(): void {
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.removeItem(this.TOKEN_KEY);
+    }
+  }
+
+  isLoggedIn(): boolean {
+    return !!this.getToken();
+  }
 
   // LOGIN
   login(username: string, password: string) {
@@ -53,11 +81,13 @@ export class AuthService {
   }
 
   logout(): void {
-    localStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+    }
   }
 
   getDecodedToken(): any | null {
-    const token = localStorage.getItem('access_token');
+    const token = this.getToken();
     if (!token) return null;
 
     try {
